Fix for PR review issue: Renamed kibitzAnalysisPolicyCopy.ts to kibitzAnalysisPolicyText.ts and changed the imports.
